+++ tests/test_resend_backend.py
@@ -4,7 +4,6 @@ from datetime import date, datetime
 from decimal import Decimal
 from email.mime.base import MIMEBase
 from email.mime.image import MIMEImage
-from email.utils import formataddr
 
 from django.core import mail
 from django.core.exceptions import ImproperlyConfigured
@@ -99,72 +98,6 @@ class ResendBackendStandardEmailTests(ResendBackendMockAPITestCase):
         self.assertEqual(
             data["bcc"], ["Blind Copy <bcc1@example.com>", "bcc2@example.com"]
         )
-
-    def test_display_name_workarounds(self):
-        # Resend's API has a bug that rejects a display-name in double quotes
-        # (per RFC 5322 section 3.4). Attempting to omit the quotes works, unless
-        # the display-name also contains a comma. Try to avoid the whole problem
-        # by using RFC 2047 encoded words for addresses Resend will parse incorrectly.
-        msg = mail.EmailMessage(
-            "Subject",
-            "Message",
-            formataddr(("Félix Företag, Inc.", "from@example.com")),
-            [
-                '"To, comma" <to1@example.com>',
-                "non–ascii <to2@example.com>",
-                "=?utf-8?q?pre_encoded?= <to3@example.com>",
-            ],
-            reply_to=['"Reply, comma" <reply1@example.com>'],
-        )
-        msg.send()
-        data = self.get_api_call_json()
-        self.assertEqual(
-            data["from"],
-            # for `from` field only, avoid RFC 2047 and retain non-ASCII characters:
-            '"Félix Företag, Inc." <from@example.com>',
-        )
-        self.assertEqual(
-            data["to"],
-            [
-                "=?utf-8?q?To=2C_comma?= <to1@example.com>",
-                "=?utf-8?b?bm9u4oCTYXNjaWk=?= <to2@example.com>",
-                "=?utf-8?q?pre_encoded?= <to3@example.com>",
-            ],
-        )
-        self.assertEqual(
-            data["reply_to"], ["=?utf-8?q?Reply=2C_comma?= <reply1@example.com>"]
-        )
-
-    @override_settings(ANYMAIL_RESEND_WORKAROUND_DISPLAY_NAME_BUGS=False)
-    def test_undocumented_workaround_setting(self):
-        # Same test as above, but workarounds disabled
-        msg = mail.EmailMessage(
-            "Subject",
-            "Message",
-            '"Félix Företag" <from@example.com>',
-            [
-                '"To, comma" <to1@example.com>',
-                "non–ascii <to2@example.com>",
-                "=?utf-8?q?pre_encoded?= <to3@example.com>",
-            ],
-            reply_to=['"Reply, comma" <reply1@example.com>'],
-        )
-        msg.send()
-        data = self.get_api_call_json()
-        self.assertEqual(
-            data["from"],
-            # (Django uses base64 encoded word unless QP is shorter)
-            "=?utf-8?b?RsOpbGl4IEbDtnJldGFn?= <from@example.com>",
-        )
-        self.assertEqual(
-            data["to"],
-            [
-                '"To, comma" <to1@example.com>',
-                "=?utf-8?b?bm9u4oCTYXNjaWk=?= <to2@example.com>",
-                "=?utf-8?q?pre_encoded?= <to3@example.com>",
-            ],
-        )
-        self.assertEqual(data["reply_to"], ['"Reply, comma" <reply1@example.com>'])
 
     def test_email_message(self):
         email = mail.EmailMessage(